DjangoRestApi/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
drf_yasg==1.21.4
